Switch to scikit-build build system
This version drops `setup.py` and so we are force to migrate to the
scikit-build build. This version of the build appears to build the main
OpenEXR library statically and link it all into the `OpenEXR.$SHLIB_EXT`
binary Python module, which means that we are *not* actually reusing the
libOpenEXR provided by the conda-forge `openexr` package. But we are
depending on conda-forge's libImath, so at least that's something.

With this new build system, we should probably turn `openexr-feedstock`
into a multi-output build and retire this feedstock. AFAICT the source
file on pypi.org is identical to the source file used for the
`openexr-feedstock` build.